Added Shared Methods to Source Generator (for cleanliness)
