+++ SKSSL.Generator/SystemGenerator.cs
@@ -3,6 +3,8 @@ using System.Linq;
 using System.Text;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
+using static SKSSL.Generator.SharedMethods;
+
 // ReSharper disable BadControlBracesIndent
 
 namespace SKSSL.Generator;
@@ -35,33 +37,6 @@ public class SystemGenerator : IIncrementalGenerator
 
             spc.AddSource(SYS_REG_NAME, GenerateSystemList(list!));
         });
-    }
-
-
-    private static SystemInfo? GetSystemInfo(GeneratorAttributeSyntaxContext ctx)
-    {
-        if (ctx.TargetSymbol is not INamedTypeSymbol symbol ||
-            symbol.IsAbstract ||
-            symbol.InstanceConstructors.All(c => c.Parameters.Length > 0))
-        {
-            // Skip abstract classes or classes without parameterless constructor.
-            return null;
-        }
-
-        AttributeData attr = ctx.Attributes[0];
-
-        var name = attr.ConstructorArguments.Length > 0
-            ? attr.ConstructorArguments[0].Value as string
-            : symbol.Name;
-
-        var priority = 0;
-        // TODO: Add support for named arguments (Priority, etc.) if needed
-
-        return new SystemInfo(
-            FullTypeName: symbol.ToDisplayString(SymbolDisplayFormat.FullyQualifiedFormat),
-            Name: name ?? symbol.Name,
-            Priority: priority
-        );
     }
 
     private static string GenerateSystemList(ImmutableArray<SystemInfo> systems)
